+++ docs/samples/storage/hf/hf_isvc.yaml
@@ -10,9 +10,9 @@ spec:
         name: huggingface
       args:
       - --model_name=llama2
-      - --model_dir=hf://meta-llama/Llama-2-7b-chat-hf
-      - --tensor_input_names=input_ids
-      - --disable_vllm
+      - --model_dir=/mnt/models
+      - --backend=huggingface
+      storageUri: hf://meta-llama/Llama-2-7b-chat-hf
       resources:
         limits:
           cpu: "6"
